backend_api/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
scikit-learn==1.4.1.post1
joblib==1.4.2
tensorflow==2.15.0
Pillow==10.3.0
scipy==1.12.0
